% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{clean_enrollment_data}
\alias{clean_enrollment_data}
\title{clean_enrollment_data}
\usage{
clean_enrollment_data(raw_enrollment_data, source = "csv")
}
\arguments{
\item{enrollment_data}{data.frame of raw enrollment data}
}
\value{
data.frame of imported, and cleaned, enrollment data
}
\description{
clean_enrollment_data
}
