% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{combine_duplicated_sections}
\alias{combine_duplicated_sections}
\title{combine_duplicated_sections}
\usage{
combine_duplicated_sections(df)
}
\arguments{
\item{df}{}
}
\description{
combine_duplicated_sections
}
