% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{format_csv_columns}
\alias{format_csv_columns}
\title{format_csv_columns}
\usage{
format_csv_columns(df)

format_csv_columns(df)
}
\arguments{
\item{df}{}
}
\description{
format_csv_columns

format_feather_columns
}
