% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_enrollment.R
\name{get_latest_low_enrolled_classes}
\alias{get_latest_low_enrolled_classes}
\title{get_latest_low_enrolled_classes}
\usage{
get_latest_low_enrolled_classes(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
data.frame only classes below the minimum enrollment
}
\description{
get_latest_low_enrolled_classes
}
