% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_enrollment_data}
\alias{read_enrollment_data}
\title{read_enrollment_data}
\usage{
read_enrollment_data(path, file_pattern)
}
\arguments{
\item{file_pattern}{}
}
\value{
data.frame of imported, and cleaned, enrollment data
}
\description{
read_enrollment_data
}
