% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_enrollment_data_from_web}
\alias{read_enrollment_data_from_web}
\title{read_enrollment_data_from_web}
\usage{
read_enrollment_data_from_web(manifest_file)
}
\arguments{
\item{manifest_file}{URL}
}
\value{
data.frame of imported, and cleaned, enrollment data
}
\description{
read_enrollment_data_from_web
}
