% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_pre_processed_enrollment_data}
\alias{read_pre_processed_enrollment_data}
\title{read_pre_processed_enrollment_data}
\usage{
read_pre_processed_enrollment_data(file_name)
}
\arguments{
\item{file_name}{}
}
\value{
data.frame of imported, and cleaned, enrollment data from feather
}
\description{
read_pre_processed_enrollment_data
}
