% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_UWL_SR_DAC_SUBJECT_IR_ALL_reports_locally}
\alias{read_UWL_SR_DAC_SUBJECT_IR_ALL_reports_locally}
\title{read_UWL_SR_DAC_SUBJECT_IR_ALL_reports_locally}
\usage{
read_UWL_SR_DAC_SUBJECT_IR_ALL_reports_locally(path, file_pattern)
}
\arguments{
\item{file_pattern}{}
}
\value{
data.frame of all files read in path of file_pattern format
}
\description{
read_UWL_SR_DAC_SUBJECT_IR_ALL_reports_locally
}
