% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_UWL_SR_DAC_SUBJECT_IR_ALL_reports_URL}
\alias{read_UWL_SR_DAC_SUBJECT_IR_ALL_reports_URL}
\title{read_UWL_SR_DAC_SUBJECT_IR_ALL_reports_URL}
\usage{
read_UWL_SR_DAC_SUBJECT_IR_ALL_reports_URL(manifest_file)
}
\arguments{
\item{manifest_file}{URL of manifest file}
}
\value{
data.frame of all files read in manifest_file
}
\description{
read_UWL_SR_DAC_SUBJECT_IR_ALL_reports_URL
}
