% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{remove_excluded_programs_and_components}
\alias{remove_excluded_programs_and_components}
\title{remove_excluded_programs_and_components}
\usage{
remove_excluded_programs_and_components(df)
}
\arguments{
\item{df}{}
}
\description{
remove_excluded_programs_and_components
}
