% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_enrollment.R
\name{determine_minimum_enrollment}
\alias{determine_minimum_enrollment}
\title{determine_minimum_enrollment}
\usage{
determine_minimum_enrollment(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
data.frame original data.frame with added column min.enrl
}
\description{
determine_minimum_enrollment
}
